requirements.txt
PySpark==3.5.1
pandas==2.2.2
numpy==1.26.4
pyarrow==15.0.2
apache-airflow==2.9.1
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.4.2
pytest==8.2.1
pytest-cov==4.1.0
python-dotenv>=1.0.0
apache-airflow-providers-papermill
dotenv-linter


